refactor: split and rename stories for Line and Rect
