+++ stories/src/line.stories.tsx
@@ -1,45 +1,226 @@
 import * as React from 'react'
 import {storiesOf} from '@storybook/react'
-import {withKnobs, number, select, boolean, text} from '@storybook/addon-knobs'
+import {boolean, number, select, text, withKnobs} from '@storybook/addon-knobs'
 
-import {Config, Plot, timeFormatter, fromFlux} from '../../giraffe/src'
+import {Config, Plot, fromFlux, timeFormatter} from '../../giraffe/src'
 import {
   binaryPrefixFormatter,
   siPrefixFormatter,
 } from '../../giraffe/src/utils/formatters'
+import {stackedLineTable} from './data/stackedLineLayer'
 import {getRandomTable} from '../../giraffe/src/utils/fixtures/randomTable'
 
 import {
   PlotContainer,
+  colorMappingKnob,
   colorSchemeKnob,
   fillKnob,
   findStringColumns,
+  getCPUTable,
   interpolationKnob,
   legendFontKnob,
   showAxesKnob,
   tickFontKnob,
   timeZoneKnob,
   tooltipColorizeRowsKnob,
+  tooltipHideKnob,
   tooltipOrientationThresholdKnob,
   xKnob,
   xScaleKnob,
   yKnob,
   yScaleKnob,
-  colorMappingKnob,
 } from './helpers'
 
-import {tooltipFalsyValues, fluxCSVAirData} from './data/fluxCSV'
+import {fluxCSVAirData, tooltipFalsyValues} from './data/fluxCSV'
 import {
-  formattableNumbersThousands,
-  formattableNumbersMillions,
   formattableNumbersBillions,
+  formattableNumbersMillions,
+  formattableNumbersThousands,
   formattableNumbersTrillions,
 } from './data/formatterCSV'
 
 const maxValue = Math.random() * Math.floor(200)
 
-storiesOf('Line Graph', module)
+storiesOf('Line', module)
   .addDecorator(withKnobs)
+  .add('Line', () => {
+    const table = getCPUTable()
+    const colors = colorSchemeKnob()
+    const legendFont = legendFontKnob()
+    const tickFont = tickFontKnob()
+    const x = xKnob(table)
+    const y = yKnob(table)
+    const xScale = xScaleKnob()
+    const yScale = yScaleKnob()
+    const timeZone = timeZoneKnob()
+    const timeFormat = select(
+      'Time Format',
+      {
+        'DD/MM/YYYY HH:mm:ss.sss': 'DD/MM/YYYY HH:mm:ss.sss',
+        'MM/DD/YYYY HH:mm:ss.sss': 'MM/DD/YYYY HH:mm:ss.sss',
+        'YYYY/MM/DD HH:mm:ss': 'YYYY/MM/DD HH:mm:ss',
+        'YYYY-MM-DD HH:mm:ss ZZ': 'YYYY-MM-DD HH:mm:ss ZZ',
+        'hh:mm a': 'hh:mm a',
+        'HH:mm': 'HH:mm',
+        'HH:mm:ss': 'HH:mm:ss',
+        'HH:mm:ss ZZ': 'HH:mm:ss ZZ',
+        'HH:mm:ss.sss': 'HH:mm:ss.sss',
+        'MMMM D, YYYY HH:mm:ss': 'MMMM D, YYYY HH:mm:ss',
+        'dddd, MMMM D, YYYY HH:mm:ss': 'dddd, MMMM D, YYYY HH:mm:ss',
+      },
+      'YYYY-MM-DD HH:mm:ss ZZ'
+    )
+    const fill = fillKnob(table, ['cpu'])
+    const interpolation = interpolationKnob()
+    const showAxes = showAxesKnob()
+    const lineWidth = number('Line Width', 1)
+    const shadeBelow = boolean('Shade Area', false)
+    const shadeBelowOpacity = number('Area Opacity', 0.1)
+    const hoverDimension = select(
+      'Hover Dimension',
+      {auto: 'auto', x: 'x', y: 'y', xy: 'xy'},
+      'auto'
+    )
+    const legendOpacity = number('Legend Opacity', 1.0, {
+      range: true,
+      min: 0,
+      max: 1.0,
+      step: 0.05,
+    })
+    const legendOrientationThreshold = tooltipOrientationThresholdKnob()
+    const legendColorizeRows = tooltipColorizeRowsKnob()
+    const legendHide = tooltipHideKnob()
+
+    const config: Config = {
+      table,
+      valueFormatters: {
+        _time: timeFormatter({timeZone, format: timeFormat}),
+        _value: val => `${Math.round(val)}%`,
+      },
+      legendFont,
+      legendOpacity,
+      legendOrientationThreshold,
+      legendColorizeRows,
+      legendHide,
+      tickFont,
+      showAxes,
+      xScale,
+      yScale,
+      layers: [
+        {
+          type: 'line',
+          x,
+          y,
+          fill,
+          interpolation,
+          colors,
+          lineWidth,
+          hoverDimension,
+          shadeBelow,
+          shadeBelowOpacity,
+        },
+      ],
+    }
+
+    return (
+      <PlotContainer>
+        <Plot config={config} />
+      </PlotContainer>
+    )
+  })
+  .add('Stacked Line Layer', () => {
+    const table = stackedLineTable
+    const colors = colorSchemeKnob()
+    const legendFont = legendFontKnob()
+    const tickFont = tickFontKnob()
+    const x = xKnob(table)
+    const y = yKnob(table)
+    const valueAxisLabel = text('Value Axis Label', 'foo')
+    const xScale = xScaleKnob()
+    const yScale = yScaleKnob()
+    const timeZone = timeZoneKnob()
+    const timeFormat = select(
+      'Time Format',
+      {
+        'DD/MM/YYYY HH:mm:ss.sss': 'DD/MM/YYYY HH:mm:ss.sss',
+        'MM/DD/YYYY HH:mm:ss.sss': 'MM/DD/YYYY HH:mm:ss.sss',
+        'YYYY/MM/DD HH:mm:ss': 'YYYY/MM/DD HH:mm:ss',
+        'YYYY-MM-DD HH:mm:ss ZZ': 'YYYY-MM-DD HH:mm:ss ZZ',
+        'hh:mm a': 'hh:mm a',
+        'HH:mm': 'HH:mm',
+        'HH:mm:ss': 'HH:mm:ss',
+        'HH:mm:ss ZZ': 'HH:mm:ss ZZ',
+        'HH:mm:ss.sss': 'HH:mm:ss.sss',
+        'MMMM D, YYYY HH:mm:ss': 'MMMM D, YYYY HH:mm:ss',
+        'dddd, MMMM D, YYYY HH:mm:ss': 'dddd, MMMM D, YYYY HH:mm:ss',
+      },
+      'YYYY-MM-DD HH:mm:ss ZZ'
+    )
+    const fill = fillKnob(table, ['cpu'])
+    const position = select(
+      'Line Position',
+      {stacked: 'stacked', overlaid: 'overlaid'},
+      'stacked'
+    )
+    const interpolation = interpolationKnob()
+    const showAxes = showAxesKnob()
+    const lineWidth = number('Line Width', 1)
+    const shadeBelow = boolean('Shade Area', false)
+    const shadeBelowOpacity = number('Area Opacity', 0.1)
+    const hoverDimension = select(
+      'Hover Dimension',
+      {auto: 'auto', x: 'x', y: 'y', xy: 'xy'},
+      'auto'
+    )
+    const legendOpacity = number('Legend Opacity', 1.0, {
+      range: true,
+      min: 0,
+      max: 1.0,
+      step: 0.05,
+    })
+    const legendOrientationThreshold = tooltipOrientationThresholdKnob()
+    const legendColorizeRows = tooltipColorizeRowsKnob()
+
+    const config: Config = {
+      table,
+      valueFormatters: {
+        _time: timeFormatter({timeZone, format: timeFormat}),
+        _value: val =>
+          `${val.toFixed(2)}${
+            valueAxisLabel ? ` ${valueAxisLabel}` : valueAxisLabel
+          }`,
+      },
+      xScale,
+      yScale,
+      legendFont,
+      tickFont,
+      showAxes,
+      legendOpacity,
+      legendOrientationThreshold,
+      legendColorizeRows,
+      layers: [
+        {
+          type: 'line',
+          x,
+          y,
+          fill,
+          position,
+          interpolation,
+          colors,
+          lineWidth,
+          hoverDimension,
+          shadeBelow,
+          shadeBelowOpacity,
+        },
+      ],
+    }
+
+    return (
+      <PlotContainer>
+        <Plot config={config} />
+      </PlotContainer>
+    )
+  })
   .add('y domain controlled mode', () => {
     const lines = number('Number of graph lines', 4)
     const fillColumnsCount = number('Number of fill columns', 5)
